LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/default-src/manifest-v2-default-src-unsafe-inline-blocked-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'default-src' contains an invalid source: ''unsafe-inline''. It will be ignored.
CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' appears in neither the script-src directive nor the default-src directive of the Content Security Policy.
This test passes if it doesn't alert fail.
